refactor(core): Moved persistense directory into infrastructure
