+++ module/core/src/Infrastructure/Persistence/Query/Builder/DbalDefaultLabelQueryBuilder.php
@@ -7,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Core\Persistence\Dbal\Query\Builder;
+namespace Ergonode\Core\Infrastructure\Persistence\Query\Builder;
 
 use Doctrine\DBAL\Query\QueryBuilder;
 use Ergonode\Core\Domain\Query\Builder\DefaultLabelQueryBuilderInterface;
