+++ module/core/src/Infrastructure/Persistence/Query/DbalLanguageTreeQuery.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * Copyright © Bold Brand Commerce Sp. z o.o. All rights reserved.
  * See LICENSE.txt for license details.
@@ -6,7 +7,7 @@
 
 declare(strict_types = 1);
 
-namespace Ergonode\Core\Persistence\Dbal\Query;
+namespace Ergonode\Core\Infrastructure\Persistence\Query;
 
 use Doctrine\DBAL\Connection;
 use Ergonode\Core\Domain\Query\LanguageTreeQueryInterface;
