+++ module/core/src/Resources/config/services.yml
@@ -4,9 +4,6 @@ services:
         autoconfigure: true
         public: false
 
-    Ergonode\Core\Persistence\:
-        resource: '../../Persistence/*'
-
     Ergonode\Core\Application\:
         resource: '../../Application/*'
 
@@ -25,8 +22,8 @@ services:
         resource: '../../Infrastructure/Handler/*'
         tags: ['messenger.message_handler']
 
-    Ergonode\Core\Persistence\Dbal\Projector\:
-        resource: '../../Persistence/Dbal/Projector/*'
+    Ergonode\Core\Infrastructure\Persistence\Projector\:
+        resource: '../../Infrastructure/Persistence/Projector/*'
         tags: ['messenger.message_handler']
 
     Ergonode\Core\Infrastructure\JMS\Serializer\Handler\:
@@ -41,14 +38,14 @@ services:
         resource: '../../Application/Form/Extension/*'
         tags: ['liform.extension']
 
-    Ergonode\Core\Domain\Query\LanguageQueryInterface: '@Ergonode\Core\Persistence\Dbal\Query\DbalLanguageQuery'
-    Ergonode\Core\Domain\Query\UnitQueryInterface: '@Ergonode\Core\Persistence\Dbal\Query\DbalUnitQuery'
-    Ergonode\Core\Domain\Repository\LanguageRepositoryInterface: '@Ergonode\Core\Persistence\Dbal\Repository\DbalLanguageRepository'
-    Ergonode\Core\Domain\Repository\UnitRepositoryInterface: '@Ergonode\Core\Persistence\Dbal\Repository\DbalUnitRepository'
+    Ergonode\Core\Domain\Query\LanguageQueryInterface: '@Ergonode\Core\Infrastructure\Persistence\Query\DbalLanguageQuery'
+    Ergonode\Core\Domain\Query\UnitQueryInterface: '@Ergonode\Core\Infrastructure\Persistence\Query\DbalUnitQuery'
+    Ergonode\Core\Domain\Repository\LanguageRepositoryInterface: '@Ergonode\Core\Infrastructure\Persistence\Repository\DbalLanguageRepository'
+    Ergonode\Core\Domain\Repository\UnitRepositoryInterface: '@Ergonode\Core\Infrastructure\Persistence\Repository\DbalUnitRepository'
     Ergonode\Core\Application\Provider\DirectoryProviderInterface: '@Ergonode\Core\Application\Provider\SimpleDirectoryProvider'
     Ergonode\Core\Infrastructure\Resolver\RelationshipsResolverInterface: '@Ergonode\Core\Infrastructure\Resolver\RelationshipsResolver'
-    Ergonode\Core\Domain\Query\Builder\DefaultLabelQueryBuilderInterface: '@Ergonode\Core\Persistence\Dbal\Query\Builder\DbalDefaultLabelQueryBuilder'
-    Ergonode\Core\Domain\Query\Builder\DefaultImageQueryBuilderInterface: '@Ergonode\Core\Persistence\Dbal\Query\Builder\DbalDefaultImageQueryBuilder'
+    Ergonode\Core\Domain\Query\Builder\DefaultLabelQueryBuilderInterface: '@Ergonode\Core\Infrastructure\Persistence\Query\Builder\DbalDefaultLabelQueryBuilder'
+    Ergonode\Core\Domain\Query\Builder\DefaultImageQueryBuilderInterface: '@Ergonode\Core\Infrastructure\Persistence\Query\Builder\DbalDefaultImageQueryBuilder'
 
     Ergonode\Core\Infrastructure\JMS\Serializer\HandlerRegistry:
         decorates: 'jms_serializer.handler_registry'
